% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exports-data.R
\docType{data}
\name{exports}
\alias{exports}
\title{UK Trade in Goods (Exports) Dataset}
\format{
\subsection{\code{exports}}{

A data frame with 226 observations and 445 variables:
\describe{
\item{columns}{Export target series (9) and monthly indicators (436).}
\item{rows}{Monthly values from Jan 2004 to Oct 2022.}
...
}
}
}
\source{
\url{https://www.ons.gov.uk/}
}
\usage{
exports
}
\description{
A full dataset used for nowcasting UK trade in goods (Exports) including the 9 export target series and 436 monthly indicator series.
}
\examples{
# load exports data 
data = exports 

}
\keyword{datasets}
