% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fillNA.R
\name{fillNA}
\alias{fillNA}
\title{Interpolation of missing data}
\usage{
fillNA(X)
}
\arguments{
\item{X}{n x p numeric matrix of stationary and standardized time series}
}
\value{
X \eqn{n \times p}{n x p} numeric matrix with missing data interpolated

idx.na \eqn{n \times p}{n x p} logical matrix with \code{TRUE} if missing and \code{FALSE} otherwise.
}
\description{
Internal missing data is filled in using a cubic spline.
Start and end of sample missing data is filled in using the median of the series and then
smoothed with an MA(3) process.
}
