% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R, R/inflation-data.R
\docType{data}
\name{inflation}
\alias{inflation}
\title{UK Inflation Dataset}
\format{
\subsection{\code{inflation}}{

A data frame with 135 observations and 36 variables:
\describe{
\item{columns}{Different classes of inflation index}
\item{rows}{Quarterly values of the relevant CPI index, benchmarked to 2015=100}
...
}
}

\subsection{\code{inflation}}{

A data frame with 135 observations and 36 variables:
\describe{
\item{columns}{Different classes of inflation index}
\item{rows}{Quarterly values of the relevant CPI index, benchmarked to 2015=100}
...
}
}
}
\source{
\url{https://www.ons.gov.uk/economy/inflationandpriceindices/datasets/consumerpriceindices}

\url{https://www.ons.gov.uk/economy/inflationandpriceindices/datasets/consumerpriceindices}
}
\usage{
inflation

inflation
}
\description{
A subset of quarterly CPI Index data from the ONS Inflation data (Q4 2022 release).

A subset of quarterly CPI Index data from the ONS Inflation data (Q4 2022 release).
}
\examples{
data = inflation 

# load inflation data 
data = inflation 

}
\keyword{datasets}
