% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{kalmanUnivariate}
\alias{kalmanUnivariate}
\title{Univariate filtering (sequential processing) for fast KFS}
\usage{
kalmanUnivariate(X, a0_0, P0_0, A, Lambda, Sig_e, Sig_u)
}
\arguments{
\item{X}{n x p, numeric matrix of (stationary) time series}

\item{a0_0}{k x 1, initial state mean vector}

\item{P0_0}{k x k, initial state covariance matrix}

\item{A}{k x k, state transition matrix}

\item{Lambda}{p x k, measurement matrix}

\item{Sig_e}{p x p, measurement equation residuals covariance matrix (diagonal)}

\item{Sig_u}{k x k, state equation residuals covariance matrix}
}
\description{
Univariate filtering (sequential processing) for fast KFS
}
