% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logspace.R
\name{logspace}
\alias{logspace}
\title{logspace}
\usage{
logspace(x1, x2, n = 50)
}
\arguments{
\item{x1}{lower bound}

\item{x2}{upper bound}

\item{n}{length}
}
\value{
Vector of log10 spaced values of length n
}
\description{
Produce a vector of log10 space values
}
