% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/outputs.R
\name{missing_data_plot}
\alias{missing_data_plot}
\title{Plot the missing data in a data matrix/frame}
\usage{
missing_data_plot(data, present.colour = "grey80", missing.colour = "grey20")
}
\arguments{
\item{data}{Numeric matrix or data frame with NA for missing values.}

\item{present.colour}{The colour for data that is present. Default is 'grey80'.}

\item{missing.colour}{The colour for data that is missing. Default is 'grey20'.}
}
\description{
Visualise the amount of missing data in a data matrix or data frame.
}
