% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/raggedEdge.R
\name{raggedEdge}
\alias{raggedEdge}
\title{Generate a ragged edge structure for a data matrix}
\usage{
raggedEdge(X, lags)
}
\arguments{
\item{X}{numeric data matrix}

\item{lags}{vector of integers representing publication lag of each variable}
}
\value{
ragged edge version of X
}
\description{
Generate a ragged edge structure for a data matrix
}
\examples{
data = matrix(rnorm(100),ncol=10)
pub_lags = c(rep(2,5),rep(1,3),rep(0,2))
new_data = raggedEdge(data, pub_lags)


}
