% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SparseDFM.R
\name{SparseDFM}
\alias{SparseDFM}
\title{Estimate a Sparse Dynamic Factor Model}
\usage{
SparseDFM(
  X,
  r,
  q = 0,
  alphas = logspace(-2, 3, 100),
  alg = "EM-sparse",
  err = "IID",
  kalman = "univariate",
  standardize = TRUE,
  max_iter = 100,
  threshold = 1e-04
)
}
\arguments{
\item{X}{A \code{n x p} numeric data matrix or data frame of (stationary) time series.}

\item{r}{Integer. Number of factors.}

\item{q}{Integer. The first q series (columns of X) should not be made sparse. Default q = 0.}

\item{alphas}{Numeric vector or value of LASSO regularisation parameters. Default is alphas = logspace(-2,3,100).}

\item{alg}{Character. Option for estimation algorithm. Default is \code{"EM-sparse"}. Options are:
\tabular{llll}{
\code{"PCA"} \tab\tab principle components analysis (PCA) for static factors seen in Stock and Watson (2002). \cr\cr
\code{"2Stage"} \tab\tab the two-stage framework of PCA plus Kalman filter/smoother seen in Giannone et al. (2008) and Doz et al. (2011). \cr\cr
\code{"EM"} \tab\tab the quasi-maximum likelihood approach using the EM algorithm to handle arbitrary patterns of missing data seen in Banbura and Modugno (2014). \cr\cr
\code{"EM-sparse"} \tab\tab the novel sparse EM approach allowing LASSO regularisation on factor loadings seen in (cite our paper). \cr\cr
}}

\item{err}{Character. Option for idiosyncratic errors. Default is \code{"IID"}. Options are:
\tabular{llll}{
\code{"IID"} \tab\tab errors are IID white noise. \cr\cr
\code{"AR1"} \tab\tab errors follow an AR(1) process. \cr\cr
}}

\item{kalman}{Character. Option for Kalman filter and smoother equations. Default is \code{"univariate"}. Options are:
\tabular{llll}{
\code{"multivariate"} \tab\tab classic Kalman filter and smoother equations seen in Shumway and Stoffer (1982). \cr\cr
\code{"univaraite"} \tab\tab univariate treatment (sequential processing) of the multivariate equations for fast Kalman filter and smoother seen in Koopman and Durbin (2000). \cr\cr
}}

\item{standardize}{Logical. Standardize the data before estimating the model. Default is \code{TRUE}.}

\item{max_iter}{Integer. Maximum number of EM iterations. Default is 100.}

\item{threshold}{Numeric. Tolerance on EM iterates. Default is 1e-4.}
}
\value{
item to be returned
}
\description{
Main function to allow estimation of a standard DFM or a sparse DFM (with sparse loadings) with:
an option for IID or AR(1) idiosyncratic errors
an option for Kalman Filter/Smoother estimation using standard multivariate equations or fast univariate filtering equations
}
\details{
Add details of the model here.
}
