% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/outputs.R
\name{summary.SparseDFM}
\alias{summary.SparseDFM}
\alias{print.SparseDFM}
\title{SparseDFM Summary Outputs}
\usage{
\method{print}{SparseDFM}(x, ...)

\method{summary}{SparseDFM}(object, ...)
}
\arguments{
\item{x}{an object of class 'SparseDFM'}

\item{\dots}{Further \code{summary} arguments.}

\item{object}{an object of class 'SparseDFM'}
}
\value{
Information on the model fitted.

Summary information on estimation details.
}
\description{
Summary and print outputs for class 'SparseDFM'.
}
