% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tune_factors.R
\name{tune_factors}
\alias{tune_factors}
\title{Tune the number of factors to use}
\usage{
tune_factors(
  X,
  type = 2,
  standardize = TRUE,
  r.max = min(15, ncol(X) - 1),
  plot = FALSE
)
}
\arguments{
\item{X}{A \code{n x p} numeric data matrix or data frame of (stationary) time series.}

\item{type}{Character. Option for which information criteria to use. Default is 2.}

\item{standardize}{Logical. Standardize the data before estimating the model. Default is \code{TRUE}.}

\item{r.max}{Integer. Maximum number of factors to search for. Default is min(15,ncol(X)-1).}

\item{plot}{Logical. Make a plot showing the IC value for each of the number of factors considered. Default is \code{FALSE}.}
}
\description{
Uses Bai and Ng (2002) informaion criteria approach. Missing data is interpolated using fill_NA() function.
}
