% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/outputs.R
\name{missing_data_plot}
\alias{missing_data_plot}
\title{Plot the missing data in a data matrix/frame}
\usage{
missing_data_plot(
  data,
  present.colour = "grey80",
  missing.colour = "grey20",
  use.names = TRUE
)
}
\arguments{
\item{data}{Numeric matrix or data frame with NA for missing values.}

\item{present.colour}{The colour for data that is present. Default is 'grey80'.}

\item{missing.colour}{The colour for data that is missing. Default is 'grey20'.}

\item{use.names}{Logical. Label the axis with data variables names. Default is TRUE. Set to FALSE to remove.}
}
\value{
A matrix plot showing where missing data is present.
}
\description{
Visualise the amount of missing data in a data matrix or data frame.
}
