% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/outputs.R
\name{residuals.SparseDFM}
\alias{residuals.SparseDFM}
\alias{fitted.SparseDFM}
\alias{resid.SparseDFM}
\title{SparseDFM Residuals and Fitted Values}
\usage{
\method{fitted}{SparseDFM}(object, standardize = FALSE, ...)

\method{residuals}{SparseDFM}(object, standardize = FALSE, ...)
}
\arguments{
\item{object}{an object of class 'SparseDFM'.}

\item{standardize}{logical. The residuals and fitted values should be standardized. Default is \code{FALSE}, values returned in the original data \eqn{\bm{X}}{X} scale.}

\item{\dots}{Further \code{residuals} arguments.}
}
\value{
Residuals or fitted values of \code{SparseDFM}.
}
\description{
Obtain the residuals or fitted values of the \code{SparseDFM} fit.
}
