% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SparseDFM.R
\docType{package}
\name{SparseDFM}
\alias{SparseDFM}
\title{SparseDFM: A package for computating the notorious bar statistic.}
\description{
Estimate a Sparse Dynamic Factor Model  (see doc for details)
}
\details{
Main function to allow estimation of a standard DFM and a sparse DFM
\itemize{
\item option for IID or AR(1) errors
\item option for Kalman Filter/Smoother estimation using standard multivariate equations or fast univariate filtering equations
}

X_t = Lambda.tilde\emph{F.tilde_t + eta_t,  eta_t ~ N(0,Sigma.eta)
F.tilde_t = A.tilde}F.tilde_{t-1} + u.tilde_t,  u.tilde_t ~ N(0,Sigma.u.tilde)

If IID errors:
\itemize{
\item Latent state of dim n x r: F.tilde_t = F_t for r factors F_t, t = 1,...,n.
\item Loadings of dim p x r: Lambda.tilde = Lambda where Lambda are the loadings (possibly) assumed to be sparse.
\item State transition matrix of dim r x r: A.tilde = A.
\item Idio error covariance of dim p x p: Sigma.eta = Sigma_epsilon.
\item State equation error covariance of dim r x r: Sigma.u.tilde = Sigma_u.
}

If AR(1) errors:
\itemize{
\item k = r + p
\item Latent state of dim n x k: F.tilde_t = \link{F_t e_t}^T for r factors F_t and p idio errors e_t., t = 1,...,n.
\item Loadings of dim k x r: Lambda.tilde = \link{Lambda I_p} where Lambda are the loadings assumed to be sparse.
\item State transition matrix of dim k x k: A.tilde = \link{A 0; 0 Phi}.
\item Idio error covariance of dim p x p: Sigma.eta = kappa*I_p, kappa set to 1e-4.
\item State equation error covariance of dim k x k: Sigma.u.tilde = \link{Sigma_u 0; 0 Sigma_epsilon}
Inputs:
}

X: n x p matrix of (stationary) time series.
r: integer, number of factors.
q: integer, the first q variables should not be made sparse. Default q = 0.
alphas: vector of lasso regularisation parameters. Default is alphas = logspace(-2,3,100).
alg: character, option for algorithm. 'PCA', '2Stage', 'EM' or 'EM-sparse'. Default is 'EM-sparse'.
err: character, option for idiosyncratic errors. 'AR1' or 'IID'. Default is 'AR1'.
kalman: character, option for Kalman filter and smoother equations. 'multivariate' or 'univariate'. Default is 'univariate'.
standardize: character, option to standardize data. TRUE or FALSE. Default is TRUE
max_iter: maximum number of EM iterations. Default is 100.
threshold: tolerance on EM iterates. Default is 1e-4.
}
\section{SparseDFM functions}{

The SparseDFM functions ...
}

