% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/outputs.R
\name{summary.sparseDFM}
\alias{summary.sparseDFM}
\alias{print.sparseDFM}
\title{sparseDFM Summary Outputs}
\usage{
\method{print}{sparseDFM}(x, ...)

\method{summary}{sparseDFM}(object, ...)
}
\arguments{
\item{x}{an object of class 'sparseDFM'}

\item{\dots}{Further \code{summary} arguments.}

\item{object}{an object of class 'sparseDFM'}
}
\value{
Information on the model fitted.

Summary information on estimation details.
}
\description{
Summary and print outputs for class 'sparseDFM'.
}
